Move `rpc` directory from `internal` to root level
- Move `rpc` and `rpc/proto` packages from `internal` to root level.
- Update import paths from `internal/rpc/proto` to `rpc/proto` in all relevant files.
